Fixing compile to set name properly on objects
