+++ api/py/ai/zipline/repo/extract_objects.py
@@ -4,7 +4,6 @@ import importlib.util
 import logging
 import os
 
-import ai.zipline.utils as utils
 from ai.zipline.logger import get_logger
 
 
@@ -28,6 +27,20 @@ def from_folder(root_path: str,
     return result
 
 
+def import_module_set_name(module, cls):
+    """evaluate imported modules to assign object name"""
+    for name, obj in list(module.__dict__.items()):
+        if isinstance(obj, cls):
+            # the name would be `team_name.python_script_name.[group_by_name|join_name|staging_query_name]`
+            # real world case: psx.reservation_status.v1
+            if hasattr(obj, "metaData"):
+                obj.metaData.name = module.__name__.partition(".")[2] + "." + name
+            else:
+                # StagingQuery has no metaData property
+                obj.name = module.__name__.partition(".")[2] + "." + name
+    return module
+
+
 def from_file(root_path: str,
               file_path: str,
               cls: type,
@@ -44,7 +57,12 @@ def from_file(root_path: str,
 
     # the key of result dict would be `team_name.python_script_name.[group_by_name|join_name|staging_query_name]`
     # real world case: psx.reservation_status.v1
-    utils.import_module_set_name(mod, cls)
-    result = {obj.name: obj
-              for obj in mod.__dict__.values() if isinstance(obj, cls)}
+    import_module_set_name(mod, cls)
+    result = {}
+
+    for obj in [o for o in mod.__dict__.values() if isinstance(o, cls)]:
+        if hasattr(obj, "metaData"):
+            result[obj.metaData.name] = obj
+        else:
+            result[obj.name] = obj
     return result
