+++ api/py/ai/zipline/utils.py
@@ -2,6 +2,7 @@ from collections.abc import Iterable
 from typing import List, Union
 
 import ai.zipline.api.ttypes as api
+import ai.zipline.repo.extract_objects as eo
 import gc
 import json
 import shutil
@@ -113,16 +114,6 @@ def get_columns(source: api.Source):
     return columns
 
 
-def import_module_set_name(module, cls):
-    """evaluate imported modules to assign object name"""
-    for name, obj in list(module.__dict__.items()):
-        if isinstance(obj, cls):
-            # the name would be `team_name.python_script_name.[group_by_name|join_name|staging_query_name]`
-            # real world case: psx.reservation_status.v1
-            obj.name = module.__name__.partition(".")[2] + "." + name
-    return module
-
-
 def get_mod_name_from_gc(obj, mod_prefix):
     """get an object's module information from garbage collector"""
     mod_name = None
@@ -138,5 +129,5 @@ def get_mod_name_from_gc(obj, mod_prefix):
 def get_staging_query_output_table_name(staging_query: api.StagingQuery):
     """generate output table name for staging query job"""
     staging_query_module = importlib.import_module(get_mod_name_from_gc(staging_query, "staging_queries"))
-    import_module_set_name(staging_query_module, api.StagingQuery)
+    eo.import_module_set_name(staging_query_module, api.StagingQuery)
     return staging_query.name.replace('.', '_')
